Integrated liveness with proxyfsd process
Also provide standalone liveness program to participate in the cluster
as before... though it now must claim one of the configured peers in
the ConfMap.

Note the updating of proxyfsd/mac_cluster_3_peers.conf and the
commensurate new proxyfsd/file_server_mac_3_peers.conf that is
a recipe to run multiple proxyfsd (or liveness) instances on a
single MacOS (leveraging alias's lo0 / localhost interfaces).