Added support for multiple UDP msg's in flight simultaneously
Previously, only a single UDP (packetized) msg was supported. Upon
arrival of a packet from a new msg, the prior incomplete msg would
be discarded. Further, if nobody "consumed" a completed msg from a
given peer, upon completion of a subsequent msg from that same peer,
the prior completed msg from that peer would be discarded.

With this change:
  1) A configurable number of incomplete msg's per peer is supported.
  2) There is no limit on the number of completed msgs (per peer or
     globally).